backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1
gunicorn
pillow
dj-database-url
psycopg2-binary
cloudinary
django-cloudinary-storage
resend